+++ u-sdk/src/oss/object/mod.rs
@@ -1,5 +1,7 @@
 mod basic;
+mod callback_verify_layer;
 mod multipart_upload;
 mod types_rs;
 
+pub use callback_verify_layer::*;
 pub use types_rs::*;